refactor collector, introduce base class
