+++ cst/plugins/eu.esdihumboldt.cst.functions.collector/META-INF/MANIFEST.MF
@@ -26,6 +26,7 @@ Import-Package: com.google.common.collect;version="17.0.0",
  eu.esdihumboldt.hale.common.schema.model.constraint.property.factory,
  eu.esdihumboldt.hale.common.schema.model.constraint.type,
  eu.esdihumboldt.util.groovy.builder,
+ eu.esdihumboldt.util.groovy.collector,
  javax.annotation
 Require-Bundle: eu.esdihumboldt.hale.common.align;bundle-version="3.2.0",
  groovy;bundle-version="2.3.7"
