+++ cst/plugins/eu.esdihumboldt.cst.functions.groovy.helpers/META-INF/MANIFEST.MF
@@ -11,6 +11,7 @@ Import-Package: com.google.common.base;version="17.0.0",
  de.fhg.igd.osgi.util;version="1.0.0",
  eu.esdihumboldt.hale.common.align.model,
  eu.esdihumboldt.hale.common.core,
+ eu.esdihumboldt.util.groovy.collector,
  javax.annotation,
  org.springframework.core.convert;version="3.1.1.RELEASE"
 Export-Package: eu.esdihumboldt.cst.functions.groovy.helpers,
