+++ util/plugins/eu.esdihumboldt.util.groovy/META-INF/MANIFEST.MF
@@ -3,8 +3,9 @@ Bundle-ManifestVersion: 2
 Bundle-Name: Groovy Utilities
 Bundle-SymbolicName: eu.esdihumboldt.util.groovy
 Bundle-Version: 3.4.0.qualifier
-Bundle-RequiredExecutionEnvironment: JavaSE-1.7
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: eu.esdihumboldt.util.groovy.builder,
+ eu.esdihumboldt.util.groovy.collector,
  eu.esdihumboldt.util.groovy.json,
  eu.esdihumboldt.util.groovy.paths,
  eu.esdihumboldt.util.groovy.xml
